重新设计支付宝支付

+++ src/main/java/com/shan/reservation/service/AlipayService.java
@@ -2,6 +2,8 @@ package com.shan.reservation.service;
 
 import com.alipay.api.AlipayApiException;
 
+import java.math.BigDecimal;
+
 /**
  * @author wsw
  * @Package com.gabe.mychat.controller
@@ -16,7 +18,7 @@ public interface AlipayService {
      * @param totalAmount   订单价格
      * @param subject       商品名称
      */
-    String webPagePay(String outTradeNo,Integer totalAmount,String subject) throws Exception;
+    String webPagePay(String outTradeNo,BigDecimal totalAmount,String subject) throws Exception;
 
     /**
      * app端订单支付
@@ -24,7 +26,7 @@ public interface AlipayService {
      * @param totalAmount   订单价格
      * @param subject       商品名称
      */
-    String appPagePay(String outTradeNo,Integer totalAmount,String subject) throws Exception;
+    String appPagePay(String outTradeNo, BigDecimal totalAmount, String subject) throws Exception;
 
     /**
      * 退款
